% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_adapter.R
\name{apply_ea_validation_rules}
\alias{apply_ea_validation_rules}
\title{Apply EA-specific validation and filtering rules}
\usage{
apply_ea_validation_rules(canonical_df, schema)
}
\arguments{
\item{canonical_df}{Canonical EA data.}

\item{schema}{EA peripheral schema.}
}
\value{
Filtered canonical EA data.
}
\description{
Applies rules defined in schema$validation.
}
