% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_04_blank.R
\name{calculate_blank_stats}
\alias{calculate_blank_stats}
\title{Calculate blank statistics}
\usage{
calculate_blank_stats(blank_df, n_blanks = NULL, verbose = TRUE)
}
\arguments{
\item{blank_df}{Dataframe containing ONLY blank measurements}

\item{n_blanks}{Number of blank replicates to average
If NULL, uses all rows in blank_df}

\item{verbose}{Print statistics (default: TRUE)}
}
\value{
List containing:
\itemize{
\item mean_area: mean area_or_voltage of blanks
\item sd_area: standard deviation of blank area
\item mean_delta: mean delta_value of blanks
\item sd_delta: standard deviation of blank delta
\item n_blanks: number of blanks used
}
}
\description{
Summarizes the blank measurements to determine correction factors.
Different correction methods may use different statistics.
}
\details{
These statistics are used by apply_blank_correction()
Different instruments may average blanks differently:
\itemize{
\item EA: average within a run, apply to that run
\item TC/EA: average all blanks, apply globally
\item GC/IRMS: average per compound, apply to that compound
}
}
