% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_01_clean.R
\name{clean_data}
\alias{clean_data}
\title{Clean canonical data: remove test runs, select peak/compound}
\usage{
clean_data(df, config, verbose = TRUE)
}
\description{
This function removes noise and artifacts from the dataset:
\itemize{
\item Removes test/quality control runs (linearity, on-off, etc.)
\item Optionally filters to specific peak number (EA)
\item Optionally filters to specific compound (GC/LC)
}
}
