% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_adapter.R
\name{get_config_element}
\alias{get_config_element}
\title{Get the element specified in the EA experiment configuration}
\usage{
get_config_element(config)
}
\arguments{
\item{config}{Experiment configuration loaded from YAML.}
}
\value{
A single element name.
}
\description{
Get the element specified in the EA experiment configuration
}
