% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_00_import.R
\name{get_peripheral_schema}
\alias{get_peripheral_schema}
\title{Get one peripheral schema}
\usage{
get_peripheral_schema(schemas, peripheral)
}
\arguments{
\item{schemas}{Loaded schemas from load_peripheral_schemas().}

\item{peripheral}{Name of the peripheral, such as "EA".}
}
\value{
The requested peripheral schema.
}
\description{
Retrieves a schema for a specific peripheral.
}
