% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_00_import.R
\name{load_IRMS_config}
\alias{load_IRMS_config}
\title{Load IRMS experimental configuration}
\usage{
load_IRMS_config(yaml_file_path, stds_reference_df = NULL, verbose = TRUE)
}
\arguments{
\item{yaml_file_path}{Path to the experiment configuration YAML file.}

\item{stds_reference_df}{Optional pre-loaded laboratory standards table.}

\item{verbose}{Print loading information.}
}
\value{
An object of class \code{irms_config}.
}
\description{
Loads an experiment configuration from YAML.
}
