% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_00_import.R
\name{load_peripheral_schemas}
\alias{load_peripheral_schemas}
\title{Load peripheral schemas}
\usage{
load_peripheral_schemas(path = NULL, config_dir = NULL)
}
\arguments{
\item{path}{Optional path to a schema YAML file.}

\item{config_dir}{Optional configuration directory. Used when developing
the package locally.}
}
\value{
A named list containing the peripheral schemas.
}
\description{
Loads the peripheral schema YAML file from the installed package.
}
