% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_03_mass.R
\name{prepare_mass_standards}
\alias{prepare_mass_standards}
\title{Calculate elemental amounts for standards}
\usage{
prepare_mass_standards(stds_df, element, config = NULL, verbose = TRUE)
}
\arguments{
\item{stds_df}{Standards dataframe from extract_standards().}

\item{element}{Element to calibrate (e.g. "C", "N", "S", "H").}

\item{config}{Configuration.}

\item{verbose}{Print progress.}
}
\value{
Standards dataframe with a new {element}.Amount column.
}
\description{
Calculates the amount of the selected element contained in each standard.
}
