% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_00_import.R
\name{validate_config}
\alias{validate_config}
\title{Validate an IRMS configuration}
\usage{
validate_config(config)
}
\arguments{
\item{config}{An object returned by load_IRMS_config().}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate an IRMS configuration
}
