% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_00_import.R
\name{validate_peripheral_schemas}
\alias{validate_peripheral_schemas}
\title{Validate all peripheral schemas}
\usage{
validate_peripheral_schemas(schemas)
}
\arguments{
\item{schemas}{Loaded schemas.}
}
\value{
Invisibly returns TRUE.
}
\description{
Performs generic structural validation of all loaded peripheral schemas.
}
